+++ repo/src/main/java/com/parashift/onlyoffice/Prepare.java
@@ -167,6 +167,9 @@ public class Prepare extends AbstractWebScript {
                 responseJson.put("mime", mimetypeService.getMimetype(docExt));
                 responseJson.put("demo", configManager.demoActive());
                 responseJson.put("documentType", documentType);
+                responseJson.put("user", username);
+                responseJson.put("share", util.getShareUrl());
+                responseJson.put("favorite", util.getFavouriteUrl(nodeRef));
 
                 logger.debug("Sending JSON prepare object");
                 logger.debug(responseJson.toString(3));