+++ share/src/main/resources/alfresco/site-webscripts/com/parashift/onlyoffice-edit.get.js
@@ -14,3 +14,6 @@ model.docTitle = pObj.config.document.title;
 model.config = JSON.stringify(pObj.config);
 model.demo = pObj.demo;
 model.documentType=pObj.documentType;
+model.share = pObj.share;
+model.user = pObj.user;
+model.favorite = pObj.favorite;