+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #
 ## Added
+- changing favorite flag of the file from document editor
 - add goBack url for document editor
 - creating new file from document editor
 - change favicon in editor by document type